Added prev and next pointers to neighboring BPA child nodes
Modified sorted_blocks to just be an array and not vector due to lack of need to resize

Made some elements private

Changed calculation of header_ptr from   log_ptr+log_size*e_size   to   just log_ptr+log_size. Believe and tested to make sure that
adding to a pointer works this way. Since the type of the ptr is int, adding a constant like 1 to it increments it by the size of an
int times this constant behind the scenes already.

Added ability to pass pre built bpa_array if desired

Added helper function getBlock